PR comments: Remove unused code
